This commit acts as a merge of the HV and lV branches
Opted to just copy in the most updated versions of each SchDoc to avoid
any problems.

The commit histories for each branch will just be located in
the respective remote branches if anything needs to be referenced. Based
on our workflow this didn't seem important to carry forward hence why I
opted to just copy in the correct files.

Signed-off-by: Ryan Bear <rabear@wisc.edu>